part2/CDS/requirements.txt
certifi==2019.3.9
chardet==3.0.4
cloudant==2.12.0
CouchDB==1.2
idna==2.8
lxml==4.3.2
numpy==1.16.4
pandas==0.24.2
Pillow==5.4.1
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.1
pywin32==224
requests==2.22.0
six==1.12.0
urllib3==1.25.3
xlrd==1.2.0
